dashboard/api/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0
pyyaml==6.0.1
docker==7.0.0
requests-unixsocket==0.4.0
